+++ src/types/Cart.type.ts
@@ -1,0 +1,6 @@
+export interface CreateCartType {
+  token: string;
+  product_id: string;
+  quantity: number;
+  check: boolean;
+}
